+++ pyctr/type/smdh.py
@@ -70,6 +70,19 @@ class AppTitle(NamedTuple):
     long_desc: str
     publisher: str
 
+    @classmethod
+    def from_bytes(cls, app_title_raw: bytes):
+        return cls(short_desc=app_title_raw[0:0x80].decode('utf-16le').strip('\0'),
+                   long_desc=app_title_raw[0x80:0x180].decode('utf-16le').strip('\0'),
+                   publisher=app_title_raw[0x180:0x200].decode('utf-16le').strip('\0'))
+
+    def __bytes__(self):
+        return b''.join((
+            self.short_desc.encode('utf-16le').ljust(0x80, b'\0'),
+            self.long_desc.encode('utf-16le').ljust(0x100, b'\0'),
+            self.publisher.encode('utf-16le').ljust(0x80, b'\0'),
+        ))
+
 
 class SMDHRegionLockout(NamedTuple):
     Japan: bool
@@ -257,9 +270,7 @@ class SMDH:
         names: Dict[str, AppTitle] = {}
         # due to region_names only being 12 elements, this will only process 12. the other 4 are unused.
         for app_title, region in zip((app_structs[x:x + 0x200] for x in range(0, 0x2000, 0x200)), region_names):
-            names[region] = AppTitle(app_title[0:0x80].decode('utf-16le').strip('\0'),
-                                     app_title[0x80:0x180].decode('utf-16le').strip('\0'),
-                                     app_title[0x180:0x200].decode('utf-16le').strip('\0'))
+            names[region] = AppTitle.from_bytes(app_title)
 
         icon_raw_small = smdh[0x2040:0x24C0]
         icon_raw_large = smdh[0x24C0:0x36C0]