+++ CHANGELOG.md
@@ -59,6 +59,7 @@ A new `nand` module with the `NAND` class is added to read and write to NAND ima
 * `KNOWN_BLOCKS` in `pyctr.type.config.save` was changed to have values be dicts with "flags" and "size" keys (instead of plain tuples)
 * Switch `get_*` and `set_*` to getters and setters in `pyctr.type.config.blocks
   * e.g. `username` instead of `get_username` and `set_username`
+* Add new `from_bytes` and `__bytes__` methods to `AppTitle` to load title structures (and modify `SMDH` to use this now)
 
 ## v0.6.0 - January 26, 2022
 ### Highlights